+++ app/src/index.tsx
@@ -36,3 +36,7 @@ const main = () => {
 };
 
 main();
+
+if (import.meta.hot) {
+  import.meta.hot.accept();
+}